+++ e3nn_jax/_src/reduced_tensor_product.py
@@ -20,6 +20,7 @@ def reduced_tensor_product_basis(
     *,
     epsilon: float = 1e-5,
     keep_ir: Optional[List[e3nn.Irrep]] = None,
+    max_order: Optional[int] = None,
     **irreps_dict,
 ) -> e3nn.IrrepsArray:
     r"""Reduce a tensor product of multiple irreps subject to some permutation symmetry given by a formula.
@@ -31,7 +32,7 @@ def reduced_tensor_product_basis(
 
         epsilon (float): the tolerance for the Gram-Schmidt orthogonalization. Default: ``1e-5``
         keep_ir (list of Irrep): irrep to keep in the output. Default: keep all irrep
-
+        max_order (int): the maximum polynomial order assuming the input to be order ``l``. Default: no limit
         irreps_dict (dict): the irreps of each index of the formula. For instance ``i="1x1o"``.
 
     Returns:
@@ -62,7 +63,7 @@ def reduced_tensor_product_basis(
         irreps_list = formula_or_irreps_list
         irreps_tuple = tuple(e3nn.Irreps(irreps) for irreps in irreps_list)
         formulas: FrozenSet[Tuple[int, Tuple[int, ...]]] = frozenset({(1, tuple(range(len(irreps_tuple))))})
-        return _reduced_tensor_product_basis(irreps_tuple, formulas, keep_ir, epsilon)
+        return _reduced_tensor_product_basis(irreps_tuple, formulas, keep_ir, epsilon, max_order)[0].simplify()
 
     formula = formula_or_irreps_list
     f0, perm_repr = germinate_perm_repr(formula)
@@ -93,7 +94,7 @@ def reduced_tensor_product_basis(
 
     irreps_tuple = tuple(irreps_dict[i] for i in f0)
 
-    return _reduced_tensor_product_basis(irreps_tuple, perm_repr, keep_ir, epsilon)
+    return _reduced_tensor_product_basis(irreps_tuple, perm_repr, keep_ir, epsilon, max_order)[0].simplify()
 
 
 def reduced_symmetric_tensor_product_basis(
@@ -102,6 +103,7 @@ def reduced_symmetric_tensor_product_basis(
     *,
     epsilon: float = 1e-5,
     keep_ir: Optional[List[e3nn.Irrep]] = None,
+    max_order: Optional[int] = None,
 ) -> e3nn.IrrepsArray:
     r"""Reduce a symmetric tensor product.
 
@@ -110,8 +112,7 @@ def reduced_symmetric_tensor_product_basis(
         order (int): the order of the tensor product. i.e. the number of indices.
         epsilon (float): the tolerance for the Gram-Schmidt orthogonalization. Default: ``1e-5``
         keep_ir (list of Irrep): irrep to keep in the output. Default: keep all irrep
-
-
+        max_order (int): the maximum polynomial order assuming the input to be order ``l``. Default: no limit
 
     Returns:
         IrrepsArray: The change of basis
@@ -123,7 +124,32 @@ def reduced_symmetric_tensor_product_basis(
 
     irreps = e3nn.Irreps(irreps)
     perm_repr: FrozenSet[Tuple[int, Tuple[int, ...]]] = frozenset((1, p) for p in itertools.permutations(range(order)))
-    return _reduced_tensor_product_basis(tuple([irreps] * order), perm_repr, keep_ir, epsilon)
+    return _reduced_tensor_product_basis(tuple([irreps] * order), perm_repr, keep_ir, epsilon, max_order)[0].simplify()
+
+
+def _sort(irreps_array: e3nn.IrrepsArray, orders: Tuple[int, ...]) -> Tuple[e3nn.IrrepsArray, Tuple[int, ...]]:
+    _, _, inv = irreps_array.irreps.sort()
+    irreps_array = irreps_array.sorted()
+    orders = [orders[i] for i in inv]
+    return irreps_array, tuple(orders)
+
+
+def _filter_ir(
+    irreps_array: e3nn.IrrepsArray, orders: Tuple[int, ...], keep_ir: FrozenSet[e3nn.Irrep]
+) -> Tuple[e3nn.IrrepsArray, Tuple[int, ...]]:
+    orders = [o for o, (_, ir) in zip(orders, irreps_array.irreps) if ir in keep_ir]
+    irreps_array = irreps_array.filtered(keep_ir)
+    return irreps_array, tuple(orders)
+
+
+def _filter_order(irreps_array: e3nn.IrrepsArray, orders, max_order: int) -> Tuple[e3nn.IrrepsArray, Tuple[int, ...]]:
+    irreps_array = e3nn.IrrepsArray.from_list(
+        [mul_ir for mul_ir, o in zip(irreps_array.irreps, orders) if o <= max_order],
+        [x for x, o in zip(irreps_array.list, orders) if o <= max_order],
+        irreps_array.shape[:-1],
+    )
+    orders = [o for o in orders if o <= max_order]
+    return irreps_array, tuple(orders)
 
 
 @functools.lru_cache(maxsize=None)
@@ -132,8 +158,9 @@ def _reduced_tensor_product_basis(
     perm_repr: FrozenSet[Tuple[int, Tuple[int, ...]]],
     keep_ir: Optional[FrozenSet[e3nn.Irrep]],
     epsilon: float,
-) -> e3nn.IrrepsArray:
-    dims = tuple(irps.dim for irps in irreps_tuple)
+    max_order: Optional[int] = None,
+) -> Tuple[e3nn.IrrepsArray, Tuple[int, ...]]:
+    dims = tuple(irreps.dim for irreps in irreps_tuple)
 
     bases = [
         (
@@ -142,27 +169,33 @@ def _reduced_tensor_product_basis(
                 irreps,
                 np.reshape(np.eye(irreps.dim), (1,) * i + (irreps.dim,) + (1,) * (len(irreps_tuple) - i - 1) + (irreps.dim,)),
             ),
+            tuple(ir.l for _, ir in irreps),  # order of the polynomial
         )
         for i, irreps in enumerate(irreps_tuple)
     ]
 
     while True:
         if len(bases) == 1:
-            f, b = bases[0]
+            f, b, ord = bases[0]
             assert f == frozenset(range(len(irreps_tuple)))
-            b = b if keep_ir is None else b.filtered(keep_ir)
-            return b.sorted().simplify()
+            if max_order is not None:
+                (b, ord) = _filter_order(b, ord, max_order)
+            if keep_ir is not None:
+                (b, ord) = _filter_ir(b, ord, keep_ir)
+            return _sort(b, ord)
 
         if len(bases) == 2:
-            (fa, a) = bases[0]
-            (fb, b) = bases[1]
+            (fa, a, oa) = bases[0]
+            (fb, b, ob) = bases[1]
             f = frozenset(fa | fb)
-            ab = reduce_basis_product(a, b, keep_ir, round_fn=round_to_sqrt_rational)
+            ab, ord = reduce_basis_product(a, oa, b, ob, max_order, keep_ir, round_fn=round_to_sqrt_rational)
             if len(subrepr_permutation(f, perm_repr)) == 1:
-                return ab.sorted().simplify()
+                return _sort(ab, ord)
             p = reduce_subgroup_permutation(f, perm_repr, dims)
-            ab = constrain_rotation_basis_by_permutation_basis(ab, p, epsilon=epsilon, round_fn=round_to_sqrt_rational)
-            return ab.sorted().simplify()
+            ab, ord = constrain_rotation_basis_by_permutation_basis(
+                ab, p, ord, epsilon=epsilon, round_fn=round_to_sqrt_rational
+            )
+            return _sort(ab, ord)
 
         # greedy algorithm
         min_p = np.inf
@@ -170,8 +203,8 @@ def _reduced_tensor_product_basis(
 
         for i in range(len(bases)):
             for j in range(i + 1, len(bases)):
-                (fa, _) = bases[i]
-                (fb, _) = bases[j]
+                (fa, _, _) = bases[i]
+                (fb, _, _) = bases[j]
                 f = frozenset(fa | fb)
                 p_dim = reduce_subgroup_permutation(f, perm_repr, dims, return_dim=True)
                 if p_dim < min_p:
@@ -183,9 +216,9 @@ def _reduced_tensor_product_basis(
         del bases[i]
         sub_irreps = tuple(irreps_tuple[i] for i in f)
         sub_perm_repr = subrepr_permutation(f, perm_repr)
-        ab = _reduced_tensor_product_basis(sub_irreps, sub_perm_repr, None, epsilon)
+        ab, ord = _reduced_tensor_product_basis(sub_irreps, sub_perm_repr, None, epsilon, max_order)
         ab = ab.reshape(tuple(dims[i] if i in f else 1 for i in range(len(dims))) + (-1,))
-        bases = [(f, ab)] + bases
+        bases = [(f, ab, ord)] + bases
 
 
 @functools.lru_cache(maxsize=None)
@@ -220,20 +253,28 @@ def germinate_perm_repr(formula: str) -> Tuple[str, FrozenSet[Tuple[int, Tuple[i
 
 def reduce_basis_product(
     basis1: e3nn.IrrepsArray,
+    order1: Tuple[int, ...],
     basis2: e3nn.IrrepsArray,
+    order2: Tuple[int, ...],
+    max_order: Optional[int] = None,
     filter_ir_out: Optional[List[e3nn.Irrep]] = None,
     round_fn=lambda x: x,
-) -> e3nn.IrrepsArray:
+) -> Tuple[e3nn.IrrepsArray, Tuple[int, ...]]:
     """Reduce the product of two basis."""
-    basis1 = basis1.sorted().simplify()
-    basis2 = basis2.sorted().simplify()
+    basis1, order1 = _sort(basis1, order1)
+    basis2, order2 = _sort(basis2, order2)
 
     new_irreps: List[Tuple[int, e3nn.Irrep]] = []
     new_list = []
+    new_orders = []
 
-    for (mul1, ir1), x1 in zip(basis1.irreps, basis1.list):
-        for (mul2, ir2), x2 in zip(basis2.irreps, basis2.list):
+    for (mul1, ir1), x1, o1 in zip(basis1.irreps, basis1.list, order1):
+        for (mul2, ir2), x2, o2 in zip(basis2.irreps, basis2.list, order2):
             for ir in ir1 * ir2:
+
+                if max_order is not None and o1 + o2 > max_order:
+                    continue
+
                 if filter_ir_out is not None and ir not in filter_ir_out:
                     continue
 
@@ -247,18 +288,20 @@ def reduce_basis_product(
                 x = np.reshape(x, x.shape[:-3] + (mul1 * mul2, ir.dim))
                 new_irreps.append((mul1 * mul2, ir))
                 new_list.append(x)
+                new_orders.append(o1 + o2)
 
     new = e3nn.IrrepsArray.from_list(new_irreps, new_list, np.broadcast_shapes(basis1.shape[:-1], basis2.shape[:-1]))
-    return new.sorted().simplify()
+    return _sort(new, tuple(new_orders))
 
 
 def constrain_rotation_basis_by_permutation_basis(
     rotation_basis: e3nn.IrrepsArray,
     permutation_basis: np.ndarray,
+    orders: Tuple[int, ...],
     *,
     epsilon=1e-5,
     round_fn=lambda x: x,
-) -> e3nn.IrrepsArray:
+) -> Tuple[e3nn.IrrepsArray, Tuple[int, ...]]:
     """Constrain a rotation basis by a permutation basis.
 
     Args:
@@ -274,18 +317,23 @@ def constrain_rotation_basis_by_permutation_basis(
 
     new_irreps: List[Tuple[int, e3nn.Irrep]] = []
     new_list: List[np.ndarray] = []
+    new_orders: List[int] = []
 
-    rotation_basis = rotation_basis.sorted().simplify()
-    for (mul, ir), rot_basis in zip(rotation_basis.irreps, rotation_basis.list):
+    for ir in sorted({ir for mul, ir in rotation_basis.irreps}):
+        idx = [i for i, (mul, ir_) in enumerate(rotation_basis.irreps) if ir == ir_]
+        rot_basis = np.concatenate([rotation_basis.list[i] for i in idx], axis=-2)
+        ord = np.array([orders[i] for i in idx for _ in range(rotation_basis.irreps[i].mul)])
+        mul = rot_basis.shape[-2]
         R = rot_basis[..., 0]
         R = np.reshape(R, (-1, mul)).T  # (mul, dim)
         P, _ = basis_intersection(R, perm, epsilon=epsilon, round_fn=round_fn)
 
-        if P.shape[0] > 0:
-            new_irreps.append((P.shape[0], ir))
-            new_list.append(round_fn(np.einsum("vu,...ui->...vi", P, rot_basis)))
+        for p in P:
+            new_irreps.append((1, ir))
+            new_list.append(round_fn(np.einsum("u,...ui->...i", p, rot_basis)[..., None, :]))
+            new_orders.append(int(np.max(ord * (p != 0.0))))
 
-    return e3nn.IrrepsArray.from_list(new_irreps, new_list, rotation_basis.shape[:-1])
+    return e3nn.IrrepsArray.from_list(new_irreps, new_list, rotation_basis.shape[:-1]), tuple(new_orders)
 
 
 def subrepr_permutation(